moved RepositoryTestCase, EventStoreTest, EventStoreManagementTest to Testing namespace
