+++ src/Broadway/ReadModel/Testing/RepositoryTestReadModel.php
@@ -11,7 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Broadway\ReadModel;
+namespace Broadway\ReadModel\Testing;
+
+use Broadway\ReadModel\SerializableReadModel;
 
 class RepositoryTestReadModel implements SerializableReadModel
 {
