+++ test/Broadway/EventStore/Management/InMemoryEventStoreManagementTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Broadway\EventStore\Management;
 
 use Broadway\EventStore\InMemoryEventStore;
+use Broadway\EventStore\Management\Testing\EventStoreManagementTest;
 
 class InMemoryEventStoreManagementTest extends EventStoreManagementTest
 {
